// nautilus-ionic/src/app/api/model/tab.ts
/**
 * nautilus-central-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';


export interface Tab { 
    id?: number;
    title?: string;
    path?: string;
    icon?: string;
    order?: number;
    roles?: Set<Role>;
}