// nautilus-ionic/src/app/api/model/utente.ts
/**
 * nautilus-central-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Utente { 
    id?: number;
    firstName?: string;
    lastName?: string;
    email?: string;
    password?: string;
}